move ifs types to ifs
